Created an algorithms package and put all algorithms there.
