+++ graph/src/test/scala/org/apache/spark/graph/AnalyticsSuite.scala
@@ -4,6 +4,7 @@ import org.scalatest.FunSuite
 
 import org.apache.spark.SparkContext
 import org.apache.spark.SparkContext._
+import org.apache.spark.graph.algorithms._
 import org.apache.spark.rdd._
 
 import org.apache.spark.graph.LocalSparkContext._
@@ -56,8 +57,8 @@ class AnalyticsSuite extends FunSuite with LocalSparkContext {
       val nVertices = 100
       val starGraph = GraphGenerators.starGraph(sc, nVertices).cache()
       val resetProb = 0.15
-      val prGraph1 = Analytics.pagerank(starGraph, 1, resetProb)
-      val prGraph2 = Analytics.pagerank(starGraph, 2, resetProb)
+      val prGraph1 = PageRank.run(starGraph, 1, resetProb)
+      val prGraph2 = PageRank.run(starGraph, 2, resetProb)
 
       val notMatching = prGraph1.vertices.zipJoin(prGraph2.vertices) { (vid, pr1, pr2) =>
         if (pr1 != pr2) { 1 } else { 0 }
@@ -71,7 +72,7 @@ class AnalyticsSuite extends FunSuite with LocalSparkContext {
       }
       assert(errors.sum === 0)
 
-      val prGraph3 = Analytics.deltaPagerank(starGraph, 0, resetProb)
+      val prGraph3 = PageRank.runUntillConvergence(starGraph, 0, resetProb)
       val errors2 = prGraph2.vertices.leftJoin(prGraph3.vertices){ (vid, pr1, pr2Opt) =>
         pr2Opt match {
           case Some(pr2) if(pr1 == pr2) => 0
@@ -88,8 +89,8 @@ class AnalyticsSuite extends FunSuite with LocalSparkContext {
     withSpark(new SparkContext("local", "test")) { sc =>
       val gridGraph = GraphGenerators.gridGraph(sc, 10, 10).cache()
       val resetProb = 0.15
-      val prGraph1 = Analytics.pagerank(gridGraph, 50, resetProb).cache()
-      val prGraph2 = Analytics.deltaPagerank(gridGraph, 0.0001, resetProb).cache()
+      val prGraph1 = PageRank.run(gridGraph, 50, resetProb).cache()
+      val prGraph2 = PageRank.runUntillConvergence(gridGraph, 0.0001, resetProb).cache()
       val error = prGraph1.vertices.zipJoin(prGraph2.vertices) { case (id, a, b) => (a - b) * (a - b) }
         .map { case (id, error) => error }.sum
       prGraph1.vertices.zipJoin(prGraph2.vertices) { (id, a, b) => (a, b, a-b) }.foreach(println(_))
@@ -110,7 +111,7 @@ class AnalyticsSuite extends FunSuite with LocalSparkContext {
   test("Grid Connected Components") {
     withSpark(new SparkContext("local", "test")) { sc =>
       val gridGraph = GraphGenerators.gridGraph(sc, 10, 10).cache()
-      val ccGraph = Analytics.connectedComponents(gridGraph).cache()
+      val ccGraph = ConnectedComponents.run(gridGraph).cache()
       val maxCCid = ccGraph.vertices.map { case (vid, ccId) => ccId }.sum
       assert(maxCCid === 0)
     }
@@ -120,7 +121,7 @@ class AnalyticsSuite extends FunSuite with LocalSparkContext {
   test("Reverse Grid Connected Components") {
     withSpark(new SparkContext("local", "test")) { sc =>
       val gridGraph = GraphGenerators.gridGraph(sc, 10, 10).reverse.cache()
-      val ccGraph = Analytics.connectedComponents(gridGraph).cache()
+      val ccGraph = ConnectedComponents.run(gridGraph).cache()
       val maxCCid = ccGraph.vertices.map { case (vid, ccId) => ccId }.sum
       assert(maxCCid === 0)
     }
@@ -133,7 +134,7 @@ class AnalyticsSuite extends FunSuite with LocalSparkContext {
       val chain2 = (10 until 20).map(x => (x, x+1) )
       val rawEdges = sc.parallelize(chain1 ++ chain2, 3).map { case (s,d) => (s.toLong, d.toLong) }
       val twoChains = Graph.fromEdgeTuples(rawEdges, 1.0).cache()
-      val ccGraph = Analytics.connectedComponents(twoChains).cache()
+      val ccGraph = ConnectedComponents.run(twoChains).cache()
       val vertices = ccGraph.vertices.collect()
       for ( (id, cc) <- vertices ) {
         if(id < 10) { assert(cc === 0) }
@@ -157,7 +158,7 @@ class AnalyticsSuite extends FunSuite with LocalSparkContext {
       val chain2 = (10 until 20).map(x => (x, x+1) )
       val rawEdges = sc.parallelize(chain1 ++ chain2, 3).map { case (s,d) => (s.toLong, d.toLong) }
       val twoChains = Graph.fromEdgeTuples(rawEdges, true).reverse.cache()
-      val ccGraph = Analytics.connectedComponents(twoChains).cache()
+      val ccGraph = ConnectedComponents.run(twoChains).cache()
       val vertices = ccGraph.vertices.collect
       for ( (id, cc) <- vertices ) {
         if (id < 10) {
@@ -182,7 +183,7 @@ class AnalyticsSuite extends FunSuite with LocalSparkContext {
     withSpark(new SparkContext("local", "test")) { sc =>
       val rawEdges = sc.parallelize(Array( 0L->1L, 1L->2L, 2L->0L ), 2)
       val graph = Graph.fromEdgeTuples(rawEdges, true).cache()
-      val triangleCount = Analytics.triangleCount(graph)
+      val triangleCount = TriangleCount.run(graph)
       val verts = triangleCount.vertices
       verts.collect.foreach { case (vid, count) => assert(count === 1) }
     }
@@ -194,7 +195,7 @@ class AnalyticsSuite extends FunSuite with LocalSparkContext {
         Array(0L -> -1L, -1L -> -2L, -2L -> 0L)
       val rawEdges = sc.parallelize(triangles, 2)
       val graph = Graph.fromEdgeTuples(rawEdges, true).cache()
-      val triangleCount = Analytics.triangleCount(graph)
+      val triangleCount = TriangleCount.run(graph)
       val verts = triangleCount.vertices
       verts.collect().foreach { case (vid, count) =>
         if (vid == 0) {
@@ -214,7 +215,7 @@ class AnalyticsSuite extends FunSuite with LocalSparkContext {
       val revTriangles = triangles.map { case (a,b) => (b,a) }
       val rawEdges = sc.parallelize(triangles ++ revTriangles, 2)
       val graph = Graph.fromEdgeTuples(rawEdges, true).cache()
-      val triangleCount = Analytics.triangleCount(graph)
+      val triangleCount = TriangleCount.run(graph)
       val verts = triangleCount.vertices
       verts.collect().foreach { case (vid, count) =>
         if (vid == 0) {
@@ -231,7 +232,7 @@ class AnalyticsSuite extends FunSuite with LocalSparkContext {
       val rawEdges = sc.parallelize(Array(0L -> 1L, 1L -> 2L, 2L -> 0L) ++
         Array(0L -> 1L, 1L -> 2L, 2L -> 0L), 2)
       val graph = Graph.fromEdgeTuples(rawEdges, true).cache()
-      val triangleCount = Analytics.triangleCount(graph)
+      val triangleCount = TriangleCount.run(graph)
       val verts = triangleCount.vertices
       verts.collect.foreach { case (vid, count) => assert(count === 1) }
     }
